pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alice-pdf"
version = "0.1.0"
description = "Extract tables from PDFs using Mistral OCR"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Andrea Borruso", email = "aborruso@gmail.com"}
]
keywords = ["pdf", "ocr", "table-extraction", "mistral", "csv"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "mistralai>=1.0.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
alice-pdf = "alice_pdf.cli:main"

[project.urls]
Homepage = "https://github.com/aborruso/alice-pdf"
Issues = "https://github.com/aborruso/alice-pdf/issues"
